Structural Pattern -- Composite
