+++ app/data/__init__.py
@@ -2,6 +2,8 @@
 from ..services.location.csbs import CSBSLocationService
 from ..services.location.jhu import JhuLocationService
 from ..services.location.nyt import NYTLocationService
+from ..services.location.dataSource import dataSource
+
 
 # Mapping of services to data-sources.
 DATA_SOURCES = {
@@ -10,12 +12,3 @@ DATA_SOURCES = {
     "nyt": NYTLocationService(),
 }
 
-
-def data_source(source):
-    """
-    Retrieves the provided data-source service.
-
-    :returns: The service.
-    :rtype: LocationService
-    """
-    return DATA_SOURCES.get(source.lower())
