+++ app/services/location/dataSource.py
@@ -1,0 +1,9 @@
+class dataSource(object):
+    def data_source(source):
+    """
+    Retrieves the provided data-source service.
+
+    :returns: The service.
+    :rtype: LocationService
+    """
+    return DATA_SOURCES.get(source.lower())
